added deletion only by user who created blog

+++ part4/controllers/blogs.js
@@ -4,12 +4,12 @@ const blogRouter = require('express').Router()
 const User = require('../models/user')
 const Blog = require('../models/blog')
 
-const getTokenFrom = request => {
-  const authorization = request.get('authorization')
-  if (authorization && authorization.startsWith('Bearer ')) {
-  return authorization.replace('Bearer ', '')}
-  return null
-}
+// const getTokenFrom = request => {
+//   const authorization = request.get('authorization')
+//   if (authorization && authorization.startsWith('Bearer ')) {
+//   return authorization.replace('Bearer ', '')}
+//   return null
+// }
 
 // HTTP GET request for all blogs
 blogRouter.get('/', async (request, response) =>{
@@ -32,8 +32,8 @@ response.status(201).json(blog)
 blogRouter.post('/', async (request, response) => {
   
   const body = request.body
-  const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
-    if (!decodedToken.id) {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
     return response.status(401).json({ error: 'token invalid' })  }
   const locatedUser = await User.findById(decodedToken.id)
 
@@ -86,10 +86,17 @@ blogRouter.put('/:id', async (request, response) => {
 
 // HTTP DELETE request to delete blog post
 blogRouter.delete('/:id', async (request, response) => {
-
-  await Blog.findByIdAndDelete(request.params.id)
-
-  response.status(204).end()
+  
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })  }
+  const locatedUser = await User.findById(decodedToken.id)
+  const blog = await Blog.findById(request.params.id)
+  if (locatedUser.id.toString() ===blog.user.toString())
+  {
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
+  } 
 })
 
 module.exports = blogRouter
